optimized

+++ Assets/Scripts/Item/Item.cs
@@ -3,9 +3,9 @@ using UnityEngine;
 
 abstract public class Item : MonoBehaviour
 {
-    [SerializeField] private int _value = 0;
+    [SerializeField] private float _value = 0;
 
-    public int Value => _value;
+    public float Value => _value;
 
     public event Action<Item> Collected;
